created useTasks custom hook
